+++ .eslintrc.json
@@ -1,13 +1,18 @@
 {
   "parserOptions": {
-    "ecmaVersion": 2020
+    "ecmaVersion": 11,
+    "sourceType": "module"
   },
   "env": {
     "es6": true,
     "node": true
   },
-  "plugins": ["github"],
-  "extends": ["plugin:github/recommended"],
+  "plugins": [
+    "github"
+  ],
+  "extends": [
+    "plugin:github/recommended"
+  ],
   "rules": {
     "github/no-then": "off",
     "no-console": "off"
